Refactored the services folder in the codebase
