+++ application-backend/models/agent_state.py
@@ -1,0 +1,21 @@
+from typing import TypedDict, List
+from models.job import Job
+
+class AgentState(TypedDict):
+    """
+    Represents the state of our job search agent.
+
+    Attributes:
+        resume_text: The user's resume content.
+        search_prompt: The user's specific search request (e.g., "remote backend jobs").
+        search_queries: A list of optimized search queries generated by the LLM.
+        retrieved_urls: A list of URLs found by the search tool.
+        extracted_jobs: A list of Job objects extracted from the URLs.
+        final_jobs: The final, filtered list of jobs to be presented to the user.
+    """
+    resume_text: str
+    search_prompt: str
+    search_queries: List[str]
+    retrieved_urls: List[str]
+    extracted_jobs: List[Job]
+    final_jobs: List[Job]
